+++ src/components/button/Button.js
@@ -91,6 +91,7 @@ export class ButtonComponent extends BaseComponent {
     this.createElement();
     this.createInput(this.element);
     this.addShortcut(this.buttonElement);
+    this.hook('input', this.buttonElement, this.element);
     if (this.component.leftIcon) {
       this.buttonElement.appendChild(this.ce('span', {
         class: this.component.leftIcon