+++ src/components/form/Form.js
@@ -46,7 +46,8 @@ export class FormComponent extends FormioForm {
    * Load the subform.
    */
   loadSubForm() {
-    if (this.subFormLoaded) {
+    // Only load the subform if the subform isn't loaded and the conditions apply.
+    if (this.subFormLoaded || !super.checkConditions(this.root ? this.root.data : this.data)) {
       return true;
     }
     this.subFormLoaded = true;
@@ -127,9 +128,9 @@ export class FormComponent extends FormioForm {
     return _.get(this.data, this.component.key).data;
   }
 
-  checkValidity() {
+  checkValidity(data, dirty) {
     // Maintain isolated data scope when passing root data for validity checks.
-    return super.checkValidity(this.subData);
+    return super.checkValidity(this.subData, dirty);
   }
 
   checkConditions() {
@@ -155,7 +156,10 @@ export class FormComponent extends FormioForm {
     // If we wish to submit the form on next page, then do that here.
     if (this.component.submit) {
       this.submitted = true;
-      return this.submit(true);
+      return this.submit(true).then(submission => {
+        // Set data to submission.
+        return this.data[this.component.key] = submission;
+      });
     }
     else {
       return super.beforeNext();