+++ src/formio.form.js
@@ -865,6 +865,10 @@ export default class FormioForm extends FormioComponents {
       error = {message: error};
     }
 
+    if ('details' in error) {
+      error = error.details;
+    }
+
     return this.showErrors(error);
   }