+++ src/formio.full.js
@@ -1,4 +1,7 @@
 'use strict';
+// DO NOT DELETE! THIS WILL BREAK PDF GENERATION.
+import * as polyfill from './formio.polyfill';
+
 import FormioWizard from './formio.wizard';
 import FormioPDF from './formio.pdf';
 import FormioForm from './formio.form';